refactor message layer
