+++ cloud/pkg/common/messagelayer/util.go
@@ -1,0 +1,149 @@
+/*
+Copyright 2022 The KubeEdge Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package messagelayer
+
+import (
+	"errors"
+	"fmt"
+	"strings"
+
+	"k8s.io/klog/v2"
+
+	"github.com/kubeedge/beehive/pkg/core/model"
+	"github.com/kubeedge/kubeedge/common/constants"
+	pkgutil "github.com/kubeedge/kubeedge/pkg/util"
+)
+
+const (
+	ResourceNode = "node"
+
+	ResourceNodeIDIndex       = 1
+	ResourceNamespaceIndex    = 2
+	ResourceResourceTypeIndex = 3
+	ResourceResourceNameIndex = 4
+
+	ResourceDeviceIndex   = 2
+	ResourceDeviceIDIndex = 3
+
+	ResourceDevice               = "device"
+	ResourceTypeTwinEdgeUpdated  = "twin/edge_updated"
+	ResourceTypeMembershipDetail = "membership/detail"
+)
+
+// BuildResource return a string as "beehive/pkg/core/model".Message.Router.Resource
+func BuildResource(nodeID, namespace, resourceType, resourceID string) (resource string, err error) {
+	if namespace == "" || resourceType == "" || nodeID == "" {
+		err = fmt.Errorf("required parameter are not set (node id, namespace or resource type)")
+		return
+	}
+
+	resource = fmt.Sprintf("%s%s%s%s%s%s%s", ResourceNode, constants.ResourceSep, nodeID, constants.ResourceSep, namespace, constants.ResourceSep, resourceType)
+	if resourceID != "" {
+		resource += fmt.Sprintf("%s%s", constants.ResourceSep, resourceID)
+	}
+	return
+}
+
+// getElementByIndex returns a string from "beehive/pkg/core/model".Message.Router.Resource by index
+func getElementByIndex(msg model.Message, index int) string {
+	sli := strings.Split(msg.GetResource(), constants.ResourceSep)
+	if len(sli) <= index {
+		return ""
+	}
+	return sli[index]
+}
+
+// GetNodeID from "beehive/pkg/core/model".Message.Router.Resource
+func GetNodeID(msg model.Message) (string, error) {
+	res := getElementByIndex(msg, ResourceNodeIDIndex)
+	if res == "" {
+		return "", fmt.Errorf("node id not found")
+	}
+	klog.V(4).Infof("The node id %s, %d", res, ResourceNodeIDIndex)
+	return res, nil
+}
+
+// GetNamespace from "beehive/pkg/core/model".Model.Router.Resource
+func GetNamespace(msg model.Message) (string, error) {
+	res := getElementByIndex(msg, ResourceNamespaceIndex)
+	if res == "" {
+		return "", fmt.Errorf("namespace not found")
+	}
+	klog.V(4).Infof("The namespace %s, %d", res, ResourceNamespaceIndex)
+	return res, nil
+}
+
+// GetResourceType from "beehive/pkg/core/model".Model.Router.Resource
+func GetResourceType(msg model.Message) (string, error) {
+	res := getElementByIndex(msg, ResourceResourceTypeIndex)
+	if res == "" {
+		return "", fmt.Errorf("resource type not found")
+	}
+	klog.V(4).Infof("The resource type is %s, %d", res, ResourceResourceTypeIndex)
+	return res, nil
+}
+
+// GetResourceName from "beehive/pkg/core/model".Model.Router.Resource
+func GetResourceName(msg model.Message) (string, error) {
+	res := getElementByIndex(msg, ResourceResourceNameIndex)
+	if res == "" {
+		return "", fmt.Errorf("resource name not found")
+	}
+	klog.V(4).Infof("The resource name is %s, %d", res, ResourceResourceNameIndex)
+	return res, nil
+}
+
+// BuildResourceForRouter return a string as "beehive/pkg/core/model".Message.Router.Resource
+func BuildResourceForRouter(resourceType, resourceID string) (string, error) {
+	if resourceID == "" || resourceType == "" {
+		return "", fmt.Errorf("required parameter are not set (resourceID or resource type)")
+	}
+	return pkgutil.ConcatStrings(resourceType, constants.ResourceSep, resourceID), nil
+}
+
+// BuildResourceForDevice return a string as "beehive/pkg/core/model".Message.Router.Resource
+func BuildResourceForDevice(nodeID, resourceType, resourceID string) (resource string, err error) {
+	if nodeID == "" || resourceType == "" {
+		err = fmt.Errorf("required parameter are not set (node id, namespace or resource type)")
+		return
+	}
+	resource = fmt.Sprintf("%s%s%s%s%s", ResourceNode, constants.ResourceSep, nodeID, constants.ResourceSep, resourceType)
+	if resourceID != "" {
+		resource += fmt.Sprintf("%s%s", constants.ResourceSep, resourceID)
+	}
+	return
+}
+
+// GetDeviceID returns the ID of the device
+func GetDeviceID(resource string) (string, error) {
+	res := strings.Split(resource, "/")
+	if len(res) >= ResourceDeviceIDIndex+1 && res[ResourceDeviceIndex] == ResourceDevice {
+		return res[ResourceDeviceIDIndex], nil
+	}
+	return "", errors.New("failed to get device id")
+}
+
+// GetResourceType returns the resourceType of message received from edge
+func GetResourceTypeForDevice(resource string) (string, error) {
+	if strings.Contains(resource, ResourceTypeTwinEdgeUpdated) {
+		return ResourceTypeTwinEdgeUpdated, nil
+	} else if strings.Contains(resource, ResourceTypeMembershipDetail) {
+		return ResourceTypeMembershipDetail, nil
+	}
+
+	return "", fmt.Errorf("unknown resource, found: %s", resource)
+}
