+++ cloud/pkg/devicecontroller/constants/service.go
@@ -2,11 +2,6 @@ package constants
 
 // Service level constants
 const (
-	ResourceNodeIDIndex          = 1
-	ResourceDeviceIndex          = 2
-	ResourceDeviceIDIndex        = 3
-	ResourceNode                 = "node"
-	ResourceDevice               = "device"
 	ResourceTypeTwinEdgeUpdated  = "twin/edge_updated"
 	ResourceTypeMembershipDetail = "membership/detail"
