+++ cloud/pkg/edgecontroller/constants/service.go
@@ -2,11 +2,6 @@ package constants
 
 // Service level constants
 const (
-	ResourceNodeIDIndex       = 1
-	ResourceNamespaceIndex    = 2
-	ResourceResourceTypeIndex = 3
-	ResourceResourceNameIndex = 4
-
 	ResourceNode = "node"
 
 	// Group
