--- cloud/pkg/edgecontroller/messagelayer/util.go
@@ -1,84 +1,0 @@
-package messagelayer
-
-import (
-	"fmt"
-	"strings"
-
-	"k8s.io/klog/v2"
-
-	"github.com/kubeedge/beehive/pkg/core/model"
-	controller "github.com/kubeedge/kubeedge/cloud/pkg/edgecontroller/constants"
-	"github.com/kubeedge/kubeedge/common/constants"
-	pkgutil "github.com/kubeedge/kubeedge/pkg/util"
-)
-
-// BuildResource return a string as "beehive/pkg/core/model".Message.Router.Resource
-func BuildResource(nodeID, namespace, resourceType, resourceID string) (resource string, err error) {
-	if namespace == "" || resourceType == "" || nodeID == "" {
-		err = fmt.Errorf("required parameter are not set (node id, namespace or resource type)")
-		return
-	}
-
-	resource = pkgutil.ConcatStrings(controller.ResourceNode, constants.ResourceSep, nodeID, constants.ResourceSep, namespace, constants.ResourceSep, resourceType)
-	if resourceID != "" {
-		resource += pkgutil.ConcatStrings(constants.ResourceSep, resourceID)
-	}
-	return
-}
-
-// BuildResourceForRouter return a string as "beehive/pkg/core/model".Message.Router.Resource
-func BuildResourceForRouter(resourceType, resourceID string) (string, error) {
-	if resourceID == "" || resourceType == "" {
-		return "", fmt.Errorf("required parameter are not set (resourceID or resource type)")
-	}
-	return pkgutil.ConcatStrings(resourceType, constants.ResourceSep, resourceID), nil
-}
-
-// getElementByIndex returns a string from "beehive/pkg/core/model".Message.Router.Resource by index
-func getElementByIndex(msg model.Message, index int) string {
-	sli := strings.Split(msg.GetResource(), constants.ResourceSep)
-	if len(sli) <= index {
-		return ""
-	}
-	return sli[index]
-}
-
-// GetNodeID from "beehive/pkg/core/model".Message.Router.Resource
-func GetNodeID(msg model.Message) (string, error) {
-	res := getElementByIndex(msg, controller.ResourceNodeIDIndex)
-	if res == "" {
-		return "", fmt.Errorf("node id not found")
-	}
-	klog.V(4).Infof("The node id %s, %d", res, controller.ResourceNodeIDIndex)
-	return res, nil
-}
-
-// GetNamespace from "beehive/pkg/core/model".Model.Router.Resource
-func GetNamespace(msg model.Message) (string, error) {
-	res := getElementByIndex(msg, controller.ResourceNamespaceIndex)
-	if res == "" {
-		return "", fmt.Errorf("namespace not found")
-	}
-	klog.V(4).Infof("The namespace %s, %d", res, controller.ResourceNamespaceIndex)
-	return res, nil
-}
-
-// GetResourceType from "beehive/pkg/core/model".Model.Router.Resource
-func GetResourceType(msg model.Message) (string, error) {
-	res := getElementByIndex(msg, controller.ResourceResourceTypeIndex)
-	if res == "" {
-		return "", fmt.Errorf("resource type not found")
-	}
-	klog.V(4).Infof("The resource type is %s, %d", res, controller.ResourceResourceTypeIndex)
-	return res, nil
-}
-
-// GetResourceName from "beehive/pkg/core/model".Model.Router.Resource
-func GetResourceName(msg model.Message) (string, error) {
-	res := getElementByIndex(msg, controller.ResourceResourceNameIndex)
-	if res == "" {
-		return "", fmt.Errorf("resource name not found")
-	}
-	klog.V(4).Infof("The resource name is %s, %d", res, controller.ResourceResourceNameIndex)
-	return res, nil
-}
